backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
motor==3.3.2
pymongo==4.6.0
sqlalchemy==2.0.23
mysql-connector-python==8.2.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
websockets==12.0
aiofiles==23.2.1
python-dotenv==1.0.0
httpx==0.25.2
pillow==10.1.0
openai==1.3.0
redis==5.0.1
celery==5.3.4
alembic==1.13.1